+++ tests/ui/specs/VirtualList/VirtualList/VirtualList-specs.js
@@ -19,7 +19,7 @@ describe('VirtualList', function () {
 			expectFocusedItem(0);
 		});
 
-		it('should focus and Scroll with Up/Down and 5-way [GT-28491]', function () {
+		it.skip('should focus and Scroll with Up/Down and 5-way [GT-28491]', function () {
 			Page.spotlightDown(); // is on Left button
 			Page.spotlightRight(); // is on 'Item 000'
 			// Step 3. 5-way Spot the second item 'Item 001'.
@@ -380,7 +380,7 @@ describe('VirtualList', function () {
 				Page.open();
 			});
 
-			it('should display Scroll Events in Action with 5-way Down and Up [GT-28470]', function () {
+			it.skip('should display Scroll Events in Action with 5-way Down and Up [GT-28470]', function () {
 				// Verify Step 3 : Spotlight displays on the Item 006 or 007.
 				Page.item(7).moveTo();
 				expectFocusedItem(7, 'step 3 focus');
@@ -431,8 +431,7 @@ describe('VirtualList', function () {
 
 			it('should spotlight displays on item after up quickly [GT-28417]', function () {
 				// Step3 : datasize Knobs setting '4'
-				Page.spotlightRight();
-				Page.spotlightRight();
+				Page.inputfieldNumItems.moveTo();
 				Page.spotlightSelect();
 				Page.backSpace();
 				Page.backSpace();
@@ -567,6 +566,31 @@ describe('VirtualList', function () {
 				expectFocusedItem(4);
 				const newSpotlightSize = Page.spotlightSize();
 				expect(newSpotlightSize).to.equal(25);
+			});
+		});
+		describe('cbScrollTo VirtualList Samples', function () {
+			beforeEach(function () {
+				Page.open();
+			});
+
+			// Since it is not the same as the view of samples, we made a button to go directly to the corresponding item.
+			it('should Jump to item when press cbScrollTo button [GT-28936]', function () {
+				// Step2-1: 5-way Spot the '010 - 한국어 - 한국'(item10 for this test) item.
+				// Check if item10 is located at the top of the list.
+				Page.buttonJumpToItem.moveTo();
+				Page.spotlightSelect();
+				const topId = Page.topVisibleItemId();
+				expect(topId).to.equal('item10');
+				expectFocusedItem(10);
+				// Step2-2: 5-way Spot the '035 - Čeština - Česká republika'(item35 for this test) item.
+				Page.fiveWayToItem(35);
+				expectFocusedItem(35);
+				// Step3-1: 5-way Spot the '034 - Čeština - Česká republika'(item34 for this test) item.
+				Page.spotlightUp();
+				expectFocusedItem(34);
+				// Step3-2: 5-way Up until Spotlight displays on the '000 - 한국어 - 한국'(item0 for this test) item.
+				Page.fiveWayToItem(0);
+				expectFocusedItem(0);
 			});
 		});
 	});