+++ CHANGELOG.md
@@ -6,17 +6,22 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 ### Added
 
+- `sandstone/Dropdown` prop `title` to optionally display a heading above the component
 - `sandstone/TooltipDecorator` prop `tooltipType` to support new transparent label-style tooltips
 - `sandstone/FixedPopupPanels` and `sandstone/FlexiblePopupPanels` prop `fullHeight` to force these components to always stretch to the screen edges
 
 ### Changed
 
+- `sandstone/Dropdown` prop `title` to `placeholder` to display a value within the component when no selection has been made
 - `sandstone/Picker` to support updated accessibility string on `joined`
 
 ### Fixed
 
+- `sandstone/ContextualPopupDecorator` to position itself correctly when `direction` is changed
+- `sandstone/FixedPopupPanels` to support accessibility properly
 - `sandstone/FixedPopupPanels` to flex to the content size and invoke scrolling (when using `sandstone/Scroller`) when the content is too big
 - `sandstone/TabLayout` performance when focusing items in the layout
+- `sandstone/Input` to update `invalidTooltip` to the latest design
 
 ## [1.0.0-beta.6] - 2020-06-08
 
@@ -46,6 +51,7 @@ The following is a curated list of changes in the Enact sandstone module, newest
 - `sandstone/Scroller` and `sandstone/VirtualList` to show the horizontal overscroll effect properly in RTL locales
 - `sandstone/TabLayout` button sizes to match the latest designs
 - `sandstone/WizardPanels` to respect using `spotlight/SpotlightContainerDecorator.spotlightDefaultClass` to determine the default focus
+- `sandstone/ContextualMenuDecorator` to not read as alert when rendered on the screen
 
 ## [1.0.0-beta.5] - 2020-06-01